replace src with nw_src
