--- new_src/api/resources/__init__.py
@@ -1,4 +1,0 @@
-from .resources import resources
-from .register import register_resources
-
-__all__ = ["resources", "register_resources"]
