--- new_src/api/resources/resources.py
@@ -1,6 +1,0 @@
-from new_src.api.resources.types import Resource
-
-
-resources_definitions = []
-
-resources = [Resource(**resource) for resource in resources_definitions]
