--- new_src/api/resources/types.py
@@ -1,11 +1,0 @@
-from dataclasses import dataclass
-from typing import Callable
-
-from new_src.api.types import MCPConcept
-
-
-@dataclass(kw_only=True)
-class Resource(MCPConcept):
-    deprecated: bool
-    func: Callable
-    uri: str
