--- new_src/api/tools/__init__.py
@@ -1,4 +1,0 @@
-from .tools import tools
-from .registery import register_tools
-
-__all__ = ["tools", "register_tools"]
