--- new_src/config/config.py
@@ -1,107 +1,0 @@
-from typing import List
-import requests
-
-from abc import ABC
-from contextvars import ContextVar
-from enum import Enum
-from pydantic import AnyHttpUrl
-from pydantic_settings import BaseSettings, SettingsConfigDict
-
-
-class Transport(str, Enum):
-    STDIO = "stdio"
-    SSE = "sse"
-    HTTP = "streamable-http"
-
-
-class Settings(ABC, BaseSettings):
-    host: str = "localhost"
-    port: int = 8000
-    s2_api_base_url: str = "https://api.singlestore.com"
-    graphql_public_endpoint: str = "https://backend.singlestore.com/public"
-    transport: Transport = Transport.STDIO
-
-
-class LocalSettings(Settings):
-    api_key: str
-    transport: Transport = Transport.STDIO
-    is_remote: bool = False
-
-    model_config = SettingsConfigDict(env_prefix="MCP_", env_file=".env.local")
-
-
-class RemoteSettings(Settings):
-    host: str
-
-    org_id: str
-
-    is_remote: bool = True
-
-    issuer_url: str
-    required_scopes: List[str]
-
-    server_url: AnyHttpUrl | None = None
-
-    client_id: str
-    callback_path: AnyHttpUrl | None = None
-
-    # SingleStore OAuth URLs
-    singlestore_auth_url: str | None = None
-    singlestore_token_url: str | None = None
-
-    # Stores temporarily generated code verifier for PKCE. Will be deleted after use.
-    singlestore_code_verifier: str = ""
-
-    model_config = SettingsConfigDict(env_prefix="MCP_", env_file=".env.remote")
-
-    def __init__(self, **data):
-        """Initialize settings with values from environment variables."""
-        super().__init__(**data)
-
-        self.server_url = AnyHttpUrl(f"http://{self.host}:{self.port}")
-        self.callback_path = AnyHttpUrl(f"http://{self.host}:{self.port}/callback")
-
-        self.singlestore_auth_url, self.singlestore_token_url = (
-            self.discover_oauth_server()
-        )
-
-    def discover_oauth_server(self) -> tuple[str, str]:
-        """Discover OAuth server endpoints"""
-        discovery_url = f"{self.issuer_url}/.well-known/openid-configuration"
-        response = requests.get(discovery_url, timeout=10)
-        response.raise_for_status()
-
-        authorization_endpoint: str = response.json().get("authorization_endpoint")
-
-        if not authorization_endpoint:
-            raise ValueError("Failed to discover OAuth endpoints")
-
-        token_endpoint: str = response.json().get("token_endpoint")
-        if not token_endpoint:
-            raise ValueError("Failed to discover OAuth endpoints")
-
-        return authorization_endpoint, token_endpoint
-
-
-_settings_ctx: ContextVar[Settings] = ContextVar("settings", default=None)
-
-
-def init_settings(transport: Transport, api_key: str | None = None) -> None:
-    match transport:
-        case Transport.HTTP:
-            _settings_ctx.set(RemoteSettings(transport=Transport.HTTP))
-        case Transport.SSE:
-            _settings_ctx.set(RemoteSettings(transport=Transport.SSE))
-        case Transport.STDIO:
-            if not api_key:
-                raise ValueError("API key must be provided for stdio transport")
-            _settings_ctx.set(LocalSettings(api_key=api_key))
-        case _:
-            raise ValueError(f"Unsupported transport mode: {transport}")
-
-
-def get_settings() -> Settings:
-    settings = _settings_ctx.get()
-    if settings is None:
-        raise RuntimeError("Settings have not been initialized.")
-    return settings
