--- src/__init__.py
@@ -1,1 +1,0 @@
-"""SingleStore MCP Server with OAuth"""
