+++ src/api/resources/resources.py
@@ -1,22 +1,6 @@
-from dataclasses import dataclass
-
-
-@dataclass
-class Resource:
-    name: str
-    description: str
-    func: callable
-    uri: str
+from src.api.resources.types import Resource
 
 
 resources_definitions = []
 
-resources = [
-    Resource(
-        name=resource["name"],
-        description=resource["description"],
-        func=resource["func"],
-        uri=resource["uri"],
-    )
-    for resource in resources_definitions
-]
+resources = [Resource(**resource) for resource in resources_definitions]
