+++ src/api/resources/types.py
@@ -1,10 +1,11 @@
 from dataclasses import dataclass
 from typing import Callable
 
-from new_src.api.types import MCPConcept
+from src.api.types import MCPConcept
 
 
 @dataclass(kw_only=True)
-class Tool(MCPConcept):
+class Resource(MCPConcept):
     deprecated: bool
     func: Callable
+    uri: str
