+++ src/api/tools/types.py
@@ -1,8 +1,10 @@
 from dataclasses import dataclass
+from typing import Callable
+
+from src.api.types import MCPConcept
 
 
-@dataclass
-class Tool:
-    name: str
-    description: str
-    func: callable
+@dataclass(kw_only=True)
+class Tool(MCPConcept):
+    deprecated: bool
+    func: Callable
