--- src/auth/__init__.py
@@ -1,3 +1,0 @@
-from .provider import SingleStoreOAuthProvider
-
-__all__ = ["SingleStoreOAuthProvider"]
