--- src/utils/logger.py
@@ -1,4 +1,0 @@
-import logging
-
-
-logger = logging.getLogger(__name__)
